hollow out the logging module
