+++ rust/gui-client/src-common/Cargo.toml
@@ -13,6 +13,7 @@ firezone-headless-client = { path = "../../headless-client" }
 firezone-logging = { workspace = true }
 hex = "0.4.3"
 minidumper = "0.8.3"
+output_vt100 = "0.1"
 rand = "0.8.5"
 reqwest = { version = "0.12.5", default-features = false, features = ["stream", "rustls-tls"] }
 secrecy = { workspace = true }
@@ -24,7 +25,10 @@ time = { version = "0.3.36", features = ["formatting"] }
 thiserror = { version = "1.0", default-features = false }
 tokio = { workspace = true }
 tracing = { workspace = true }
+tracing-log = "0.2"
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 url = { version = "2.5.2" }
+zip = { version = "2", features = ["deflate", "time"], default-features = false }
 
 [dependencies.keyring]
 version = "3.2.1"
