+++ rust/gui-client/src-common/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod auth;
 pub mod crash_handling;
 pub mod deep_link;
+pub mod logging;
 pub mod updates;
