+++ rust/gui-client/src-tauri/src/client/gui/errors.rs
@@ -1,5 +1,5 @@
-use super::{deep_link, logging};
 use anyhow::Result;
+use firezone_gui_client_common::{self as common, deep_link};
 use firezone_headless_client::{ipc, IpcServiceError, FIREZONE_GROUP};
 
 // TODO: Replace with `anyhow` gradually per <https://github.com/firezone/firezone/pull/3546#discussion_r1477114789>
@@ -11,7 +11,7 @@ pub(crate) enum Error {
     #[error("Deep-link module error: {0}")]
     DeepLink(#[from] deep_link::Error),
     #[error("Logging module error: {0}")]
-    Logging(#[from] logging::Error),
+    Logging(#[from] common::logging::Error),
 
     // `client.rs` provides a more user-friendly message when showing the error dialog box for certain variants
     #[error("IPC")]
